ref #25 - the dreaded file split and fixing phpcs warnings/errors
