+++ .gitignore
@@ -3,3 +3,4 @@
 .sass-cache
 /sass/
 /compass/
+/vendor/
