+++ phpunit.xml.dist
@@ -11,7 +11,8 @@
     convertWarningsToExceptions = "true"
     processIsolation            = "false"
     stopOnFailure               = "false"
-    syntaxCheck                 = "false">
+    syntaxCheck                 = "false"
+    bootstrap                   = "vendor/autoload.php">
 
     <testsuites>
         <testsuite name="Project Test Suite">
