+++ src/Formatter/Compressed.php
@@ -1,0 +1,58 @@
+<?php
+/**
+ * SCSSPHP
+ *
+ * @copyright 2012-2014 Leaf Corcoran
+ *
+ * @license http://opensource.org/licenses/gpl-license GPL-3.0
+ * @license http://opensource.org/licenses/MIT MIT
+ *
+ * @link http://leafo.net/scssphp
+ */
+
+namespace Leafo\ScssPhp\Formatter;
+
+use Leafo\ScssPhp\Formatter;
+
+/**
+ * SCSS compressed formatter
+ *
+ * @author Leaf Corcoran <leafot@gmail.com>
+ */
+class Compressed extends Formatter
+{
+    public function __construct()
+    {
+        $this->indentLevel = 0;
+        $this->indentChar = '  ';
+        $this->break = '';
+        $this->open = '{';
+        $this->close = '}';
+        $this->tagSeparator = ',';
+        $this->assignSeparator = ':';
+    }
+
+    public function indentStr($n = 0)
+    {
+        return '';
+    }
+
+    public function blockLines($inner, $block)
+    {
+        $glue = $this->break.$inner;
+
+        foreach ($block->lines as $index => $line) {
+            if (substr($line, 0, 2) === '/*' && substr($line, 2, 1) !== '!') {
+                unset($block->lines[$index]);
+            } elseif (substr($line, 0, 3) === '/*!') {
+                $block->lines[$index] = '/*' . substr($line, 3);
+            }
+        }
+
+        echo $inner . implode($glue, $block->lines);
+
+        if (!empty($block->children)) {
+            echo $this->break;
+        }
+    }
+}
